WHAT THE FUCK!!!!!!!!!!
You need to manually install aws-sdk in your project to get sqlite3 to work. I
never read this anywhere and, at least because React is trying to be fancy by
clearin the screen every so often, the actual error message that would've
informed me of this is only available momentarily.

Let me reiterate,

WHAT! THE! FUCK!!!!!!!!!!!!!